Make PPA code Position Independent
This patch zeroizes BSS section, turns on PIE and adds support
for R_AARCH64_RELATIVE relocations. R_AARCH64_RELATIVE represents
a relative adjustment to the place based on the load address of
the object relative to its original link address.  At the time of
linking it is not known where PPA will be loaded.  So the load
address needs to be dynamically calculated and R_AARCH64_RELATIVE
type relocations need to be taken care of.

RELA style relocation has the following relocation structure
	src addr, info (Relocation Type), addend

Pseudo alg used for relocations:
  if info == R_AARCH64_RELATIVE
	reloc_addr = src_addr - link_addr + load_addr
	reloc value = addend - link_addr + load_addr
	*reloc_addr = reloc_value

Signed-off-by: Ruchika Gupta <ruchika.gupta@nxp.com>
Signed-off-by: unknown <B44983@B44983-11.fsl.freescale.net>