driver/sfp: changes otpmk random bytes dest. addr. from OCRAM to DDR.
On SoC's LS1088A, LS2088A the OCRAM is default marked cacheable and
is connected to HNI port. Cache coherancy is not maintained for OCRAM
and hence the core is not able to see the data written by CAAM on OCRAM.

Changes the random bytes destination address to be used by CAAM from
OCRAM to heap section in DDR. Uses alloc method to get the memory and
that address is passed as destination address.

Signed-off-by: Udit Agarwal <udit.agarwal@nxp.com>
Reviewed-by: Ruchika Gupta <ruchika.gupta@nxp.com>
Signed-off-by: Rod Dorris <rod.dorris@nxp.com>